+++ index.html
@@ -104,7 +104,6 @@
                 <div class="col-md-12">
                     <!-- Alert -->
                      <div class="alert alert-dismissable show fade" role="alert">
-                        <button type="button" class="close" data-dismiss="alert" aria-label="Close">
                             <span aria-hidden="true"></span>
                         </button>
                         <strong></strong>
