+++ styles.css
@@ -1,80 +1,78 @@
 body,
 html {
-    height: 100%;
-    background-color: #fffffe;
-    color: #272343;
+  height: 100%;
+  background-color: #fffffe;
+  color: #272343;
 }
 p {
-    padding: 0;
-    margin: 0;
+  padding: 0;
+  margin: 0;
 }
 
-
-
 .scroll-area-sm {
-    height: 50vh;
-    overflow-x: hidden
+  height: 50vh;
+  overflow-x: hidden;
 }
 
 .dot {
-    height: 15px;
-    width: 15px;
-    display: inline-block;
+  height: 15px;
+  width: 15px;
+  display: inline-block;
 }
 
 .icon:hover {
-    opacity: 60%;
+  opacity: 60%;
 }
 
 .taskSummary {
-    display: flex;
-    list-style-type: none;
-    justify-content: space-around;
-    align-items: center;
+  display: flex;
+  list-style-type: none;
+  justify-content: space-around;
+  align-items: center;
 }
 
 .done {
-    background-color: rgba(40, 167, 69, 0.2)
+  background-color: rgba(40, 167, 69, 0.2);
 }
 
 .done-font {
-    color: rgba(40, 167, 69, 0.5);
+  color: rgba(40, 167, 69, 0.5);
 }
 
 .review {
-    background-color: rgba(23, 162, 184, 0.2)
+  background-color: rgba(23, 162, 184, 0.2);
 }
 
 .review-font {
-    color: rgba(23, 162, 184, 0.5)
+  color: rgba(23, 162, 184, 0.5);
 }
 
 .inProgress {
-    background-color: rgba(255, 193, 7, 0.2)
+  background-color: rgba(255, 193, 7, 0.2);
 }
 
 .inProgress-font {
-    color: rgba(255, 193, 7, 0.5)
+  color: rgba(255, 193, 7, 0.5);
 }
 
 .toDo {
-    background-color: rgba(220, 53, 69, 0.2)
+  background-color: rgba(220, 53, 69, 0.2);
 }
 
 .toDo-font {
-    color: rgba(220, 53, 69, 0.5)
+  color: rgba(220, 53, 69, 0.5);
 }
 .custom-checkbox {
-    padding: 0px;
+  padding: 0px;
 }
 .task-detail-summary {
-    padding-left: 0.7rem ;
+  padding-left: 0.7rem;
 }
-.alertModal-danger {
+.alert-danger {
   color: #721c24;
   background: #f8d7da;
 }
-.alertModal-success {
+.alert-success {
   color: #155724;
   background: #d4edda;
 }
